Search the PolarView catalogue

This function is used by bb_handler_polarview(). Users probably won't need to use it directly.

Usage

bb_polarview_search(
  acquisition_date = Sys.Date() + -14:0,
  formats = c("jpg", "geotiff"),
  polygon = NULL,
  max_results = 200L,
  verbose = FALSE
)

Arguments

acquisition_date

Date: the allowable image acquisition dates

formats

character: one or more of "jpg" (jpg preview images) "jp2" or "geotiff". Note that the geotiffs are much larger than the jpg previews

polygon

string or sfc_POLYGON: either an sfc_POLYGON or a string giving a polygon in WKT format and EPSG:3031 projection. Only images intersecting this polygon will be returned

max_results

integer: maximum number of results to return

verbose

logical: if TRUE, show additional progress output

Value

A character vector of image URLs

References

https://www.polarview.aq