Meteorological data sources

Data sources providing meteorological data.

Usage

sources_meteorological(name, formats, time_resolutions, ...)

Arguments

name

character vector: only return data sources with name or id matching these values

formats

character: for some sources, the format can be specified. See the list of sources above for details

time_resolutions

character: for some sources, the time resolution can be specified. See the list of sources above for details

...

: additional source-specific parameters. See the list of sources above for details

Value

a tibble with columns as specified by bowerbird::bb_source()

Details

"Antarctic Mesoscale Prediction System grib files": The Antarctic Mesoscale Prediction System - AMPS - is an experimental, real-time numerical weather prediction capability that provides support for the United States Antarctic Program, Antarctic science, and international Antarctic efforts

The returned tibble contains more information about each source.

References

See the doc_url and citation field in each row of the returned tibble for references associated with these particular data sources

Examples

if (FALSE) { # \dontrun{
## define a configuration and add the AMPS data to it
cf <- bb_config("/my/file/root") %>%
  bb_add(sources_meteorological("Antarctic Mesoscale Prediction System grib files"))
} # }