Data sources intended for species distribution modelling and similar tasks

Data sources providing environmental and similar gridded data, suitable for species distribution modelling, regionalisation analyses, and similar tasks.

Usage

sources_sdm(name, formats, time_resolutions, ...)

Arguments

name

character vector: only return data sources with name or id matching these values

formats

character: for some sources, the format can be specified. See the list of sources above for details

time_resolutions

character: for some sources, the time resolution can be specified. See the list of sources above for details

...

: additional source-specific parameters. See the list of sources above for details

Value

a tibble with columns as specified by bowerbird::bb_source()

Details

"Southern Ocean marine environmental data": a collection of gridded marine environmental data layers suitable for use in Southern Ocean species distribution modelling. All environmental layers have been generated at a spatial resolution of 0.1 degrees, covering the Southern Ocean extent (80 degrees S - 45 degrees S, -180 - 180 degrees). The layers include information relating to bathymetry, sea ice, ocean currents, primary production, particulate organic carbon, and other oceanographic data. See the vignette for more information: vignette("SO_SDM_data", package = "blueant")

The returned tibble contains more information about each source.

References

See the doc_url and citation field in each row of the returned tibble for references associated with these particular data sources

Examples

if (FALSE) { # \dontrun{
## define a configuration, storing data in a temporary directory
cf <- bb_config(local_file_root = tempdir())

## add the marine environmental data layers
cf <- cf %>% bb_add(sources_sdm("Southern Ocean marine environmental data"))

## sync it (get the data)
res <- bb_sync(cf, verbose = TRUE)

## see the vignette for more information on this data source:
vignette("SO_SDM_data", package = "blueant")
} # }