Topographical data sources

Data sources providing topographical data.

sources_topography(name, formats, time_resolutions, ...)

Arguments

name
character vector: only return data sources with name or id matching these values

formats
character: for some sources, the format can be specified. See the list of sources above for details

time_resolutions
character: for some sources, the time resolution can be specified. See the list of sources above for details

...
: additional source-specific parameters. See the list of sources above for details

Value

a tibble with columns as specified by bb_source

Details

"Smith and Sandwell bathymetry": Global seafloor topography from satellite altimetry and ship depth soundings

"ETOPO1 bathymetry": ETOPO1 is a 1 arc-minute global relief model of Earth's surface that integrates land topography and ocean bathymetry

"ETOPO2 bathymetry": 2-Minute Gridded Global Relief Data (ETOPO2v2c)

"Bedmap2": Bedmap2 is a suite of gridded products describing surface elevation, ice-thickness and the sea floor and subglacial bed elevation of the Antarctic south of 60S

"Kerguelen Plateau bathymetric grid 2010": digital elevation model (DEM) for the Kerguelen Plateau region

"George V bathymetry": Digital Elevation Models (DEMs) of varying resolutions for the George V and Terre Adelie continental margin, derived by incorporating all available singlebeam and multibeam point depth data

"Geoscience Australia multibeam bathymetric grids of the Macquarie Ridge": This is a compilation of all the processed multibeam bathymetry data that are publicly available in Geoscience Australia's data holding for the Macquarie Ridge

"IBCSO bathymetry": The International Bathymetric Chart of the Southern Ocean (IBCSO) Version 1.0 is a digital bathymetric model portraying the seafloor of the circum-Antarctic waters south of 60S. IBCSO Version 1.0 has been compiled from all available bathymetric data collectively gathered by more than 30 institutions from 15 countries, including multibeam and single-beam echo soundings, digitized depths from nautical charts, regional bathymetric gridded compilations, and predicted bathymetry

"IBCSO chart for printing": The IBCSO Poster, 2013, is a polar stereographic view of the Southern Ocean displaying bathymetric contours south of 60S at a scale of 1:7,000,000

"RTOPO-1 Antarctic ice shelf topography": a consistent dataset of Antarctic ice sheet topography, cavity geometry, and global bathymetry

"Radarsat Antarctic digital elevation model V2": The high-resolution Radarsat Antarctic Mapping Project (RAMP) digital elevation model (DEM) combines topographic data from a variety of sources to provide consistent coverage of all of Antarctica. Version 2 improves upon the original version by incorporating new topographic data, error corrections, extended coverage, and other modifications

"New Zealand Regional Bathymetry 2016": The NZ 250m gridded bathymetric data set and imagery, Mitchell et al. 2012, released 2016

"Cryosat-2 digital elevation model": a digital elevation model of Antarctica derived from 6 years of continuous CryoSat-2 measurements

"Natural Earth 10m physical vector data": Natural Earth is a public domain map dataset available at 1:10m, 1:50m, and 1:110 million scales

"GSHHG coastline data": a Global Self-consistent, Hierarchical, High-resolution Geography Database

"Shuttle Radar Topography Mission elevation data SRTMGL1 V3": Global 1-arc-second topographic data generated from NASA's Shuttle Radar Topography Mission. Version 3.0 (aka SRTM Plus or Void Filled) removes all of the void areas by incorporating data from other sources such as the ASTER GDEM

"Reference Elevation Model of Antarctica mosaic tiles": The Reference Elevation Model of Antarctica (REMA) is a high resolution, time-stamped digital surface model of Antarctica at 8-meter spatial resolution (and reduced-resolution, resampled versions). Accepts a single spatial_resolution value of "1km", "200m" [default], "100m", "8m"

The returned tibble contains more information about each source.

References

See the doc_url and citation field in each row of the returned tibble for references associated with these particular data sources

See also

sources_altimetry, sources_biological, sources_meteorological, sources_ocean_colour, sources_oceanographic, sources_reanalysis, sources_sdm, sources_seaice, sources_sst

Examples

# NOT RUN {
## define a configuration and add the Smith and Sandwell bathymetry
cf <- bb_config("/my/file/root") %>%
  bb_add(sources_topography("Smith and Sandwell bathymetry"))
# }