Reanalysis data sources

Data sources providing data from global reanalysis models.

sources_reanalysis(name, formats, time_resolutions, ...)

Arguments

name
character vector: only return data sources with name or id matching these values

formats
character: for some sources, the format can be specified. See the list of sources above for details

time_resolutions
character: for some sources, the time resolution can be specified. See the list of sources above for details

...
: additional source-specific parameters. See the list of sources above for details

Value

a tibble with columns as specified by bb_source

Details

"NCEP-DOE Reanalysis 2": NCEP-DOE Reanalysis 2 is an improved version of the NCEP Reanalysis I model that fixed errors and updated paramterizations of of physical processes. Accepts time_resolution values of "6 hour", "day", and/or "month" (default). The 6-hourly data is the original output time resolution. Daily and monthly averages are calculated from the 6-hourly model output

"NCEP/NCAR Reanalysis 1": The NCEP/NCAR Reanalysis 1 project is using a state-of-the-art analysis/forecast system to perform data assimilation using past data from 1948 to the present. Only the monthly data are so far included here

The returned tibble contains more information about each source.

References

See the doc_url and citation field in each row of the returned tibble for references associated with these particular data sources

See also

sources_altimetry, sources_meteorological, sources_ocean_colour, sources_oceanographic, sources_seaice, sources_sst, sources_topography

Examples

# NOT RUN {
## define a configuration and add the monthly NCEP2 data to it
cf <- bb_config("/my/file/root") %>%
  bb_add(sources_reanalysis("NCEP-DOE Reanalysis 2",time_resolution="month"))
# }