Sea ice data sources

Data sources providing (primarily satellite-derived) sea ice data:

"NSIDC SMMR-SSM/I Nasateam sea ice concentration": Passive microwave estimates of sea ice concentration at 25km spatial resolution. Daily and monthly resolution, available from 1-Oct-1978 to near-present. Data undergo a quality checking process and are updated annually. More recent data are available via the 'NSIDC SMMR-SSM/I Nasateam near-real-time sea ice concentration' source

"NSIDC SMMR-SSM/I Nasateam near-real-time sea ice concentration": Near-real-time passive microwave estimates of sea ice concentration at 25km, daily resolution. For older, quality-controlled data see the 'NSIDC SMMR-SSM/I Nasateam sea ice concentration' source

"NSIDC passive microwave supporting files": Grids and other support files for NSIDC passive microwave sea ice data

"Nimbus Ice Edge Points from Nimbus Visible Imagery": This data set (NmIcEdg2) estimates the location of the North and South Pole sea ice edges at various times during the mid to late 1960s, based on recovered Nimbus 1 (1964), Nimbus 2 (1966), and Nimbus 3 (1969) visible imagery

"Artist AMSR-E sea ice concentration": Passive microwave estimates of daily sea ice concentration at 6.25km spatial resolution, from 19-Jun-2002 to 2-Oct-2011. Accepts formats "geotiff" and/or "hdf"

"Artist AMSR-E supporting files": Grids and other support files for Artist AMSR-E passive microwave sea ice data

"Artist AMSR2 near-real-time sea ice concentration": Near-real-time passive microwave estimates of daily sea ice concentration at 6.25km spatial resolution, from 24-July-2012 to present

"Artist AMSR2 near-real-time 3.125km sea ice concentration": Near-real-time passive microwave estimates of daily sea ice concentration at 3.125km spatial resolution, from 1-Jan-2016 to present

"Artist AMSR2 regional sea ice concentration": Passive microwave estimates of daily sea ice concentration at 3.125km spatial resolution in selected regions, from 24-July-2012 to present

"Artist AMSR2 supporting files": Grids and landmasks for Artist AMSR2 passive microwave sea ice data

"CERSAT SSM/I sea ice concentration": Passive microwave sea ice concentration data at 12.5km resolution, 3-Dec-1991 to present

"CERSAT SSM/I sea ice concentration supporting files": Grids for the CERSAT SSM/I sea ice concentration data

"MODIS Composite Based Maps of East Antarctic Fast Ice Coverage": Maps of East Antarctic landfast sea-ice extent, generated from approx. 250,000 1 km visible/thermal infrared cloud-free MODIS composite imagery (augmented with AMSR-E 6.25-km sea-ice concentration composite imagery when required). Coverage from 2000-03-01 to 2008-12-31

"National Ice Center Antarctic daily sea ice charts": The USNIC Daily Ice Edge product depicts the daily sea ice pack in red (8-10/10ths or greater of sea ice), and the Marginal Ice Zone (MIZ) in yellow. The marginal ice zone is the transition between the open ocean (ice free) and pack ice. The MIZ is very dynamic and affects the air-ocean heat transport, as well as being a significant factor in navigational safety. The daily ice edge is analyzed by sea ice experts using multiple sources of near real time satellite data, derived satellite products, buoy data, weather, and analyst interpretation of current sea ice conditions. The product is a current depiction of the location of the ice edge vice a satellite derived ice edge product

sources_seaice(name, formats, time_resolutions, ...)

Arguments

name
character vector: only return data sources with name or id matching these values

formats
character: for some sources, the format can be specified. See the list of sources above for details

time_resolutions
character: for some sources, the time resolution can be specified. See the list of sources above for details

...
: additional source-specific parameters. See the list of sources above for details

Value

a tibble with columns as specified by bb_source

Details

The returned tibble contains more information about each source.

References

See the doc_url and citation field in each row of the returned tibble for references associated with these particular data sources

See also

sources_altimetry, sources_meteorological, sources_ocean_colour, sources_oceanographic, sources_reanalysis, sources_sst, sources_topography

Examples

# NOT RUN {
## define a configuration and add the AMSR-E data to it (geotiff format)
cf <- bb_config("/my/file/root") %>%
  bb_add(sources_seaice("Artist AMSR-E sea ice concentration",formats="geotiff"))
# }